requirements.txt
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
contourpy==1.3.1
cryptography==43.0.3
cssselect==1.2.0
cycler==0.12.1
defusedxml==0.7.1
et_xmlfile==2.0.0
fonttools==4.55.0
fpdf2==2.8.1
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdfminer.six==20240706
pdfquery==0.4.3
pillow==11.0.0
pycparser==2.22
pycparser-fake-libc==2.21
pyparsing==3.2.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
roman==4.2
six==1.16.0
tzdata==2024.2
xlrd==2.0.1